Fix cookies for learn flows
Accidentally the same cookie was being used for all learned flows, when
it should have been the cookie for the pod being addressed in the learn
action.

Additionally, change cookies for pods to be by IP, which is more useful
than a hash of pod name and namespace. Now there is a flow cache entry
per pod IP.

Signed-off-by: Tim Rozet <trozet@redhat.com>