Fix learned flows caching
Previous behavior was when we learn a flow for table 20, we ignore any
future updates. However, that is incorrect behavior, because the flows
in table 20 are learned (dst mac is dynamic), and we would miss updating
our cache with that change.

Real use case scenario is GW is floating IP and the IP floats from one
chassis to another. In this case the mac changes, and we miss updating
our flow cache with that.

To fix this, store the unique learn flow in its own field of the cache
entry, and update it anytime it changes (other than when we are
requested to ignore learning on an NS update).

Signed-off-by: Tim Rozet <trozet@redhat.com>