Fixes not sending arp response back go exgw
Due to the change to allow us to learn the dest mac to gw via any arp
packet, we were not responding someimtes to arp request from gw because
the aforementioned flow was consuming it. This patch moves the response
handling to table 2, and now either possible path ends up in table 2 to
ensure we send back an arp response.

Signed-off-by: Tim Rozet <trozet@redhat.com>