Remove idle timeout on learned flows
There's no point in having this. We cache the flows anyway and the flow
will be deleted when the pod is deleted. If the pod is still around we
re-add the flow anyway. In the case of a GW failover with a new mac, our
learn action will overwrite this flow, and we will pick up the new flow
when we resync. The idle_timeout only adds confusion.

Also includes minor cosmetic fix to the tunnel flow cookie to be the
right value.

Signed-off-by: Tim Rozet <trozet@redhat.com>