Implement routing-namespace and routing-network annotations behaviour
Allows for Pod External Gateways to be used as ECMP next hop egress
traffic. The pod must be either host networked or contain
appropriate routing-network annotation to indicate a multus network to
use. The determination of a pod's eligibility as an external gateway is
automatic and triggered by annotating the pod with the routing-namespace
annotation.

Co-authored-by: Tim Rozet <trozet@redhat.com>
Signed-off-by: Tim Rozet <trozet@redhat.com>