Pod Add: handle existing annotation errors
Currently if a pod is added with an existing annotation, we configure it
in OVN, but never try to allocate the IPs in our allocator. Potentially
this could result in another pod being created and getting the same IP
address since it is untracked by the allocator. This patch ensures the
allocator is able to get the address, otherwise clear the existing
annotation.

Signed-off-by: Tim Rozet <trozet@redhat.com>