add port: try to annotate pods before OVN commands
Since we can rely on CNI to wait until the OpenFlow flows are present,
we can safely try to annotate the pod first on pod create. Most of the
time the add_logical_port function fails because the pod has been
deleted in k8s, so by the time we fail to annotate the pod, all of the
OVN configuration has already been done.

This patch now tries to annotate first, and then execute the commands
after if the annotation is successful. IPAM is still rolled back if
annotation fails.

Signed-off-by: Tim Rozet <trozet@redhat.com>