pod -> nodePort traffic dont go to host
Using shared gw mode and a pod going to a local node port service
results in the traffic flowing to ovn_cluster_router, then back to the
node switch, out to the host where it gets SNAT/DNAT and comes back into
the node switch to hit the cluster service load balancer. Rather than
doing that, simply add the N/S load balancer that we already use in the
GR to DNAT directly to the endpoint within the node switch.

This means pod -> nodePort goes to node switch, then DNAT directly to
endpoint.

Signed-off-by: Tim Rozet <trozet@redhat.com>