pods should use EUI64 link-local address
Pods should use an EUI64 generated link-local address, because
it is being used by the MLD protocol as source address (RFC3810
5.1.14) and it will be dropped by OVN. OVN drops all traffic that
is not using a link-local address based on the mac address
of the interface.

We were replacing the mac address of the interface, but the
link-local address was the one generated from the old mac
address. Setting down and up the interface, regenerates the link-
local address with the new mac address.

Signed-off-by: Antonio Ojea <aojea@redhat.com>