Fix local node port access for hybrid mode
Previously when a node tries to access node port service, the packet is
DNAT'ed to the OVN GR and sent that way out the ovn-k8s-gw0 port.
However, the GR receives the packet with the source IP/MAC the same as
the host and the GR's hybrid port and drops it.

This patch modifies node port access from hosts to DNAT to the service
cluster IP/port so that the packet will get sent into ovn-k8s-mp0 and
get SNAT'ed.

Signed-off-by: Tim Rozet <trozet@redhat.com>