Fixes External IP with local gateway mode
For external IP usage where the IP does not exist on the host, we create
a route to get the packet into OVN where the load balancer will DNAT it.
However after the changes to new local gateway mode this packet needs be
routed to ovn_cluster_router via the mp0 interface.

Signed-off-by: Tim Rozet <trozet@redhat.com>