Fixes Hybrid Overlay for services with hybrid endpoints
With Hybrid Overlay endpoint for a service, a pod trying to reach this
service will go to the OVN switch where it will be DNAT'ed and forwarded
into ovn_cluster_router. In Local gw mode, the packet will be routed to
the kernel via mp0, then routed back into OVN towards the .3 address. In
shared gateway mode the packet will be sent towards the GR.

We need this traffic to be rerouted to the .3 address directly. This
commit adds a logical route policy to do so on ovn_cluster_router.

Signed-off-by: Tim Rozet <trozet@redhat.com>