ovn node, gw router: set node masquerade IP on br-ex
Since the `br-ex` bridge already has the node masquerade IP address, it
is no longer required to add the masquerade route to the nodes.

The SNAT/unSNAT traffic to/from the node to use the node masquerade IP
address will **still** be used, to specifically support the use case of
packets sent from the node having a different source IP address.

The `V4MasqueradeSubnet` had to be updated, since follow-up commits will
require the `.4` address on that subnet to use as a dummy next-hop for
deployments where there isn't a next-hop available.

Signed-off-by: Miguel Duarte Barroso <mdbarroso@redhat.com>
Co-authored-by: Tim Rozet <trozet@redhat.com>