Remove default gateway requirement for deployments
Now that host -> service can work without using a default gateway there
is no need to require it. With this change no next hops will be written
to the l3 gateway config annotation for the node if none are detected.
However, this only will work if --gateway-interface config is provided
at ovnkube node start time. If that is not provided, we still rely on
default gateway lookup to determine which interface should be used to
communicate with the external bridge.

Signed-off-by: Tim Rozet <trozet@redhat.com>