Fixes NAD Controller syncAll
There are several issues with NAD Controller start up. Some of the
issues were introduced by adb1fc8b5f

Issues:
 1. We were logging an error and continuing the function, instead of
    returning.
 2. For migration from annotation on node, the code was being called
    even when the allocator was nil (zone, node modes), causing panic.
 3. The same group of NADs were being accidentally sync'ed twice.
 4. For zone/node modes, syncAll would attempt to sync a NAD, and if the
    NAD did not yet have an ID, it would return nil, ignoring the NAD
    and allowing start up.

This patch changes the behavior so that if this is a zone/node
controller, we will wait up to 60 seconds for NADs to get their network
ID annotated. Additional fixes for the previous issues also included.

Unit tests added to cover the different scenarios.

Signed-off-by: Tim Rozet <trozet@nvidia.com>