Fixes NAD Controller syncAll - upgrading with networkID
There are several issues with NAD Controller start up. Some of the
issues were introduced by adb1fc8b5f

The core problem is that with the change to move the networkID from
nodes to NADs, the upgrade logic left a gap in time where pods could not
start. This because cluster-manager was responsible for migrating the
networkID from the node->NAD, and in our upgrade strategy, workers
upgrade before control plane nodes. This would leave worker nodes in a
state where new OVNK code was running that was only looking for the
networkID on the NAD, but it had not yet been migrated.

This patch changes the behavior so that any NAD Controller (zone, node,
or cluster manager) will attempt at start up to find NADs that are
missing networkIDs and search nodes for the legacy values. The NAD
Controller will then attempt to reserve the ID for the entire cluster by
updating the NAD itself.

Unit tests added to cover the different scenarios.

Signed-off-by: Tim Rozet <trozet@nvidia.com>