Fixes NAD Controller syncAll for networkID upgrade from node->NAD
Fixes issues were introduced by adb1fc8

The core problem is that with the change to move the networkID from
nodes to NADs, the upgrade logic left a gap in time where pods could not
start. This because cluster-manager was responsible for migrating the
networkID from the node->NAD, and in our upgrade strategy, workers
upgrade before control plane nodes. This would leave worker nodes in a
state where new OVNK code was running that was only looking for the
networkID on the NAD, but it had not yet been migrated.

This patch changes the behavior so that any NAD Controller (zone, node,
or cluster manager) will attempt at start up to find NADs that are
missing networkIDs and search nodes for the legacy values. Node and Zone
NAD controllers will fallback to the legacy ID, but will not annotate
the NAD. Cluster manager will also use the legacy ID and update the NAD
with it.

Unit tests added to cover the different scenarios.

Signed-off-by: Tim Rozet <trozet@nvidia.com>